requirements.txt
python-samsung-mdc
streamlit
